backend/pyproject.toml
[project]
name = "gitsummarize"
version = "0.1.0"
description = ""
authors = [
    {name = "Antarix N",email = "antarix18@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "google-genai (>=1.8.0,<2.0.0)",
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "aiohttp (>=3.11.14,<4.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "supabase (>=2.15.0,<3.0.0)",
    "openai (>=1.69.0,<2.0.0)"
]

[tool.poetry]
packages = [{include = "gitsummarize", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
